Update pushable_block_class.py
added debugging lines of code to see if everything works. everything except the player being able to collide and push works, it worked last time. i think i deleted or changed something and forgot about it. i am too deep into debugging to find out what i changed, it says it is colliding, and it is moving at the block speed from old position to new position, but it isn't and the player is going through the block. i don't think i can fix this on time